Get the “Typing React State” exercises working locally
This adds Vite to the exercises so we can develop them locally and not rely on CodeSandbox figuring out how to run our exercises.
