--- typing-react-state/Exercise1/problem/styles.css
@@ -1,3 +1,0 @@
-section {
-  margin-bottom: 10px;
-}
