--- typing-react-state/Exercise1/problem/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "include": [
-      "./**/*"
-  ],
-  "compilerOptions": {
-      "strict": true,
-      "esModuleInterop": true,
-      "lib": [
-          "dom",
-          "es2015"
-      ],
-      "jsx": "react-jsx"
-  }
-}
