+++ typing-react-state/Exercise1/solution/src/App.tsx
@@ -5,14 +5,25 @@
 //    will have to be changed after you have added the correct type to the
 //    useState hook for companyDetails.
 
-import "./styles.css";
 import React, { useState, useEffect } from "react";
 
+interface CompanyDetailProps {
+  size: number;
+  owner: string;
+  departments: string[];
+};
+
 const App = () => {
-  const [companyName, setCompanyName] = useState();
-  const [companyDescription, setCompanyDescription] = useState(null);
-  const [companyDetails, setCompanyDetails] = useState({});
-  const [employees, setEmployees] = useState([]);
+  const [companyName, setCompanyName] = useState("");
+  const [companyDescription, setCompanyDescription] = useState<string | null>(
+    null
+  );
+  const [companyDetails, setCompanyDetails] = useState<CompanyDetailProps>({
+    size: 0,
+    owner: "",
+    departments: []
+  });
+  const [employees, setEmployees] = useState<string[]>();
 
   useEffect(() => {
     setCompanyName("ABC Inc.");
@@ -58,9 +69,9 @@ const App = () => {
           <button
             onClick={() =>
               setCompanyDetails({
-                size: "",
-                owner: 0,
-                departments: 0
+                size: 0,
+                owner: "",
+                departments: []
               })
             }
           >
