--- typing-react-state/Exercise1/solution/styles.css
@@ -1,3 +1,0 @@
-section {
-  margin-bottom: 10px;
-}
