--- typing-react-state/Exercise2/problem/index.tsx
@@ -1,6 +1,0 @@
-import { render } from "react-dom";
-
-import App from "./App";
-
-const rootElement = document.getElementById("root");
-render(<App />, rootElement);
