+++ typing-react-state/Exercise2/solution/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Typing React State</title>
+    <style>
+        html {
+            color-scheme: light dark;
+            font-family: system-ui, sans-serif;
+        }
+    </style>
+</head>
+
+<body>
+    <noscript>
+        You need to enable JavaScript to run this app.
+    </noscript>
+    <main></main>
+    <script src="./src/index.tsx" type="module"></script>
+</body>
+
+</html>
