+++ typing-react-state/Exercise2/solution/package.json
@@ -1,15 +1,24 @@
 {
-  "name": "typing-react-state",
+  "name": "@bitovi/trainings-typing-react-state-use-reducer",
   "version": "1.0.0",
-  "main": "index.tsx",
+  "author": {
+    "email": "contact@bitovi.com",
+    "name": "Bitovi",
+    "url": "https://www.bitovi.com/services/training/react"
+  },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "4.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "17.0.0",
-    "@types/react-dom": "17.0.0",
-    "typescript": "4.1.3"
+    "@types/react": "^18.2.25",
+    "@types/react-dom": "^18.2.10",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.11"
+  },
+  "license": "UNLICENSED",
+  "private": true,
+  "scripts": {
+    "start": "vite"
   }
 }
