--- typing-react-state/Exercise2/solution/public/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html> 
- <html lang="en"> 
-  
- <head> 
-     <meta charset="utf-8" /> 
-     <title>Typing React State</title> 
-     <meta name="viewport" content="width=device-width, initial-scale=1.0" /> 
-     <style> 
-         html { 
-             font-family: system-ui, sans-serif; 
-         } 
-  
-         code { 
-             font-family: ui-monospace, monospace; 
-         } 
-     </style> 
- </head> 
-
-<body>
-	<noscript>
-		You need to enable JavaScript to run this app.
-	</noscript>
-	<div id="root"></div>
-</body>
-
-</html>
