+++ typing-react-state/Exercise2/solution/src/App.tsx
