--- typing-react-state/Exercise2/solution/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "include": [
-      "./**/*"
-  ],
-  "compilerOptions": {
-      "strict": true,
-      "esModuleInterop": true,
-      "lib": [
-          "dom",
-          "es2015"
-      ],
-      "jsx": "react-jsx"
-  }
-}
