+++ typing-react-state/README.md
@@ -2,17 +2,17 @@
 
 ## Exercises
 
-**Exercise 1**
+### `useState` (exercise 1)
 
-- [Problem](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise1/problem?file=/App.tsx)
-- [Solution](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise1/solution?file=/App.tsx)
+- [Problem](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise1/problem?file=src/App.tsx)
+- [Solution](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise1/solution?file=src/App.tsx)
 
-**Exercise 2**
+### `useReducer` (exercise 2)
 
-- [Problem](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise2/problem?file=/App.tsx)
-- [Solution](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise2/solution?file=/App.tsx)
+- [Problem](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise2/problem?file=src/App.tsx)
+- [Solution](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise2/solution?file=src/App.tsx)
 
-**Exercise 3**
+### `useRef` (exercise 3)
 
-- [Problem](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise3/problem?file=/App.tsx)
-- [Solution](https://codesandbox.io/s/github/bitovi/trainings/tree/main/typing-react-state/Exercise3/solution?file=/App.tsx)
+- [Problem](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise3/problem?file=src/App.tsx)
+- [Solution](https://codesandbox.io/p/sandbox/github/bitovi/trainings/tree/main/typing-react-state/Exercise3/solution?file=src/App.tsx)
